+++ Sources/Client/Client/Client.swift
@@ -181,15 +181,10 @@ public final class Client: Uploader {
     ///   - config: the configuration object with details of how the new instance should be set up.
     ///   - defaultURLSessionConfiguration: the base URLSession configuration `Client` uses for its URL sessions.
     ///                                     `Client` is allowed to override the configuration with its own settings.
-    ///   - defaultWebSocketProviderType: the default WebSocket provider type. `Client` will create it on set user.
+    ///   - defaultWebSocketProviderType: the default WebSocket provider type. `Client` will create it on set user. `config.defaultWebSocketProviderType` will override this.
     init(config: Client.Config,
          defaultURLSessionConfiguration: URLSessionConfiguration = .default,
-         defaultWebSocketProviderType: WebSocketProvider.Type = {
-        if #available(iOS 13, *) {
-            return URLSessionWebSocketProvider.self
-        }
-        return StarscreamWebSocketProvider.self
-        }()) {
+         defaultWebSocketProviderType: WebSocketProvider.Type? = nil) {
         apiKey = config.apiKey
         baseURL = config.baseURL
         callbackQueue = config.callbackQueue ?? .global(qos: .userInitiated)
@@ -199,7 +194,19 @@ public final class Client: Uploader {
         logger = logOptions.logger(icon: "🐴", for: [.requestsError, .requests, .requestsInfo])
         
         self.defaultURLSessionConfiguration = defaultURLSessionConfiguration
-        self.defaultWebSocketProviderType = defaultWebSocketProviderType
+        
+        if let defaultWebSocketProviderType = defaultWebSocketProviderType {
+            self.defaultWebSocketProviderType = defaultWebSocketProviderType
+        } else if #available(iOS 13, *) {
+            switch config.webSocketProviderType {
+            case .native:
+                self.defaultWebSocketProviderType = URLSessionWebSocketProvider.self
+            case .starscream:
+                self.defaultWebSocketProviderType = StarscreamWebSocketProvider.self
+            }
+        } else {
+            self.defaultWebSocketProviderType = StarscreamWebSocketProvider.self
+        }
         
         // Init the WebSocket to register subscriptions when the Client is initiated.
         _ = webSocket